pyproject.toml
[project]
name = "suara-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "db-dtypes>=1.5.1",
    "faster-whisper>=1.2.1",
    "google-cloud-bigquery>=3.41.0",
    "google-cloud-storage>=3.10.1",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.2",
    "kaggle>=2.0.0"
]
